Reorganize
Move 'styles' folder from /src/globals to /src/assets
Move 'constants' folder from /src/global to /src/
Remove 'globals' folder
Fix relevant imports
